--- frontend/src/components/pages/ExercisesList.css
@@ -1,10 +1,0 @@
-.exercise-list__main {
-  padding-left: 20px;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(auto-fill, 1fr);
-  grid-column-gap: 20px;
-  grid-row-gap: 20px;
-  }
-
-
