+++ frontend/src/styles/ExerciseListItem.scss
