+++ frontend/src/styles/Exercises.scss
@@ -1,5 +1,5 @@
 .exec {
   margin: 40px;
-  font-size: 20px;
+  font-size: 25px;
   font-family:Arial, Helvetica, sans-serif;
 }
